+++ src/main/kotlin/com/coder/gateway/models/WorkspaceAgentModel.kt
@@ -2,8 +2,11 @@ package com.coder.gateway.models
 
 import com.coder.gateway.sdk.Arch
 import com.coder.gateway.sdk.OS
+import java.util.UUID
 
 data class WorkspaceAgentModel(
+    val workspaceID: UUID,
+    val workspaceName: String,
     val name: String,
     val templateName: String,
     val status: WorkspaceVersionStatus,